+++ src/shared/Network/SocketMgrWin32.cpp
@@ -9,7 +9,7 @@
 #include "Network.h"
 #ifdef CONFIG_USE_IOCP
 
-#include <CrashHandler.h>
+#include "Debugging/CrashHandler.h"
 
 SocketMgr& SocketMgr::getInstance()
 {
