+++ src/world/Server/WorldRunnable.cpp
@@ -6,7 +6,7 @@ This file is released under the MIT license. See README-MIT for more information
 
 
 #include "WorldRunnable.h"
-#include <CrashHandler.h>
+#include "Debugging/CrashHandler.h"
 #include "World.h"
 #include "ServerState.h"
